fix(debug): Fix runtime error introduced by #781 (#790)
* fix(debug): Import JSON tree Svelte component to fix runtime error

Relying on the compiled JavaScript of the JSON Tree component caused 
runtime errors. The best explanation I’ve read is here: 
<https://github.com/sveltejs/svelte/issues/3671#issuecomment-544329688>

* test(debug): Mock `svelte-json-tree-auto` in tests

`svelte-json-tree-auto` causes errors in tests (see #781). I think this 
might have something to do with the fact that the tree has circular 
dependencies, which perhaps the Jest transformer can’t handle. As it’s a 
third-party library, we can probably do without it during tests, and 
this change replaces it with an empty Svelte component for tests.

* test(debug): Add interactivity test for Debug panel