feat(debug): Events shortcuts, stages support & minor fixes (#791)
* refactor(debug): Don’t assign shortcuts for events

We were assigning shortcuts for boardgame.io events (`endTurn` etc.), 
which mostly made the shortcut assignment code redundant, because the 
repeated events starting with `e` would never allow us to assign 
shortcuts matching the first letter of moves. This was additionally 
redundant as we weren’t actually using the event shortcuts anywhere.

* feat(debug): Add keyboard shortcuts for built-in events

Uses the `Move` component to render the `endStage`, `endTurn`, and 
`endPhase` events as interactive functions with associated keyboard 
shortcuts. Shortcuts are 7, 8, 9 to minimize collisions with move 
shortcuts.

Addresses #771

* fix(debug): Fix use of :global selector

Fixes #782

* fix(debug): Fix colour contrast of `null` values in JSON Tree (a11y)

* fix(debug): Fix isConnected/multiplayer logic in Info panel

* refactor(debug): Simplify global `button` and `select` styles

* feat(debug): Use buttons for player IDs to make them keyboard accessible

Addresses #771

* fix(debug): Use latest moves & events dispatchers

Closes #771

Previously, the debug panel only read the moves & events dispatchers 
from the client when instantiated. This prevented users from toggling 
between players and dispatching moves from the debug panel as different 
players with stages, because the move/event would be out-of-date 
(created with the wrong playerID). This change makes sure we’re always 
referencing the latest dispatchers.

* fix(debug): Activate interactive function when it receives focus

The content-editable field in the interactive function can receive 
focus, e.g. by tabbing to it. Activating it correctly disables other 
shortcuts during typing and styles the focus state as expected.