feat(types): loosen typing of move arguments
I had the idea of typing move arguments using ts-toolbelt’s `Misc.JSON.Value` type which represents the various possible results of parsing a JSON object because that is what is permitted for serialization over the wire. This makes it harder to type moves though because you can no longer do `({ G }, id: number) => {}`. On the one hand this could be nice as it reminds the move to narrow the type as it comes potentially from an untrusted source (similar to if typing with `unknown`) but that doesn’t seem to quite justify the additional frustrating overhead when writing moves.